flask-docker-master/requirements.txt
category_encoders==2.2.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.1
Flask-SSLify==0.1.5
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
MarkupSafe==1.1.1
netaddr==0.7.19
numpy==1.18.1
pandas==0.25.3
plotly==4.4.1
psycopg2-binary==2.8.4
pymongo==3.10.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.22.0
retrying==1.3.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.12
urllib3==1.25.8
Werkzeug==0.16.0
whitenoise==5.0.1
